cleanup: extents_from_size is 32bit value
Current lvm2 does support only 32bit extents count.

+++ lib/metadata/metadata-exported.h
@@ -731,7 +731,7 @@ int lv_rename_update(struct cmd_context *cmd, struct logical_volume *lv,
 int lv_update_and_reload(struct logical_volume *lv);
 int lv_update_and_reload_origin(struct logical_volume *lv);
 
-uint64_t extents_from_size(struct cmd_context *cmd, uint64_t size,
+uint32_t extents_from_size(struct cmd_context *cmd, uint64_t size,
 			   uint32_t extent_size);
 uint32_t extents_from_percent_size(struct volume_group *vg, const struct dm_list *pvh,
 				   uint32_t extents, int roundup,
